refactor: move media close to usage
