+++ src/App.js
@@ -1,6 +1,7 @@
 import { useCallback, useEffect, useState } from "react";
 import axios from "axios";
-import NavigationBar from "./components/NavigationBar";
+
+import { NavigationBar } from "./components/NavigationBar/NavigationBar";
 
 import "./App.css";
 
@@ -69,7 +70,7 @@ export const App = () => {
       }}
     >
       <NavigationBar />
-      
+
       <div
         // TODO: Styles can be defined in a seperate file using mui useStyle
         style={{
