+++ src/components/NavigationBar/index.js
@@ -1,0 +1,1 @@
+export {NavigationBar} from "./NavigationBar";
